move s3 temp file deletion up to handler
